+++ kernels/genweightsKernel.cu
@@ -1,7 +1,7 @@
 #include <kernelDefs.h>
 #include <thrust/random.h>
 
-__global__ void genWeightsKern( kernelArray<double> ref, uint32_t in, kernelArray<int> params, size_t device_offset){
+__global__ void genWeightsKern( kernelArray<double> ref, size_t in, kernelArray<int> params, size_t device_offset){
     size_t idx = blockIdx.x * blockDim.x + threadIdx.x;
     size_t ind = params.array[11] + idx + device_offset; // offset to start of weights, stride of num weights (stride of numWeights)
     thrust::minstd_rand0 randEng;
